+++ docs/js/elements.js
@@ -114,7 +114,7 @@ var _spritejs = spritejs,
   text2.attr({
     pos: [500, 40],
     fillColor: '#077',
-    font: '64px 宋体',
+    font: '64px "宋体"',
     lineHeight: 112,
     textAlign: 'center',
     padding: [0, 30],